+++ kgx/__init__.py
@@ -1,4 +1,4 @@
 """
 KGX Package
 """
-__version__ = "1.5.7"
+__version__ = "1.7.0"